requirements.txt
alabaster==0.7.12
aniso8601==7.0.0
Babel==2.7.0
boto3==1.9.187
botocore==1.12.187
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
docutils==0.14
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-RESTful==0.3.7
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.16.4
numpydoc==0.9.1
opencv-python==4.1.0.25
packaging==19.0
Pillow==6.1.0
pyDOE2==1.2.0
Pygments==2.4.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
smt==0.3.4
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
urllib3==1.25.3
Werkzeug==0.15.4
